requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
joblib==1.3.2
llama-index==0.10.6
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.12.0
packaging==23.2
pandas==2.2.0
pydantic==2.6.1
pydantic_core==2.16.2
pypdf==4.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
wrapt==1.16.0
yarl==1.9.4